requirements.txt
streamlit
openai-whisper
torchaudio
requests
torch
transformers
scipy